+++ example-plugin/src/main/kotlin/cat/daisy/series/example/ExampleSeriesConfig.kt
@@ -6,4 +6,9 @@ object ExampleSeriesConfig {
     val flags: List<String> = listOf("hide enchants", "hide attributes")
     val enchantment: String = "sharpness"
     val effect: String = "slow falling"
+    val biome: String = "cherry grove"
+    val entity: String = "zombie villager"
+    val gameMode: String = "surv"
+    val particle: String = "totem"
+    val statistic: String = "player kills"
 }